Port config to TypeScript
